lib: restructure teads, imports, and comments.
